+++ src/app/my-bookings/api/booking/[id]/route.ts
@@ -1,25 +1,45 @@
 import { connectDB } from "@/lib/connectDB";
 import { ObjectId } from "mongodb";
+import { NextRequest, NextResponse } from "next/server";
 
-export const DELETE = async (request, { params }) => {
+export const DELETE = async (
+  request: NextRequest,
+  { params }: { params: { id: string } }
+) => {
   const db = await connectDB();
+  if (!db) {
+    return NextResponse.json(
+      { message: "Database connection failed" },
+      { status: 500 }
+    );
+  }
   const bookingsCollection = db.collection("bookings");
   const { id } = params;
   try {
     const resp = await bookingsCollection.deleteOne({
       _id: new ObjectId(id),
     });
-    return Response.json({
-      meaasge: " delete booking product",
+    return NextResponse.json({
+      message: "Deleted booking product",
       response: resp,
     });
   } catch (error) {
-    return Response.json({ meaasge: " Somthing wrong" });
+    console.error(error); // Log the error for debugging
+    return NextResponse.json({ message: "Something went wrong" });
   }
 };
 
-export const PATCH = async (request, { params }) => {
+export const PATCH = async (
+  request: NextRequest,
+  { params }: { params: { id: string } }
+) => {
   const db = await connectDB();
+  if (!db) {
+    return NextResponse.json(
+      { message: "Database connection failed" },
+      { status: 500 }
+    );
+  }
   const bookingsCollection = db.collection("bookings");
 
   const updateDoc = await request.json();
@@ -35,17 +55,27 @@ export const PATCH = async (request, { params }) => {
         upsert: true,
       }
     );
-    return Response.json({
-      meaasge: " update booking product",
+    return NextResponse.json({
+      message: "Updated booking product",
       response: resp,
     });
   } catch (error) {
-    return Response.json({ meaasge: " Somthing wrong" });
+    console.error(error); // Log the error for debugging
+    return NextResponse.json({ message: "Something went wrong" });
   }
 };
 
-export const GET = async (request, { params }) => {
+export const GET = async (
+  request: NextRequest,
+  { params }: { params: { id: string } }
+) => {
   const db = await connectDB();
+  if (!db) {
+    return NextResponse.json(
+      { message: "Database connection failed" },
+      { status: 500 }
+    );
+  }
   const bookingsCollection = db.collection("bookings");
   const { id } = params;
 
@@ -53,11 +83,12 @@ export const GET = async (request, { params }) => {
     const resp = await bookingsCollection.findOne({
       _id: new ObjectId(id),
     });
-    return Response.json({
-      meaasge: " booking found product",
+    return NextResponse.json({
+      message: "Booking found",
       data: resp,
     });
   } catch (error) {
-    return Response.json({ meaasge: " Somthing wrong" });
+    console.error(error); // Log the error for debugging
+    return NextResponse.json({ message: "Something went wrong" });
   }
 };
